+++ bulwark/__init__.py
@@ -1,1 +1,9 @@
-name = "bulwark"
+import os
+import sys
+sys.path.insert(0, os.path.abspath('../'))
+
+import project_info
+
+
+__name__ = project_info.NAME
+__version__ = project_info.RELEASE