+++ tests/test_checks.py
@@ -62,6 +62,17 @@ def test_is_shape():
         dc.IsShape(shape=(9, 2), cheese=True)(_add_n)(df)  # bad dc param check
 
 
+def test_has_no_x():
+    df = pd.DataFrame([1, 2, 3], index=['a', 'b', 'c'])
+    result = ck.has_no_x(df, values=['x', 'y', 'z'])
+    tm.assert_frame_equal(df, result)
+
+    result = dc.HasNoX()(_add_n)(df, 2)
+    tm.assert_frame_equal(result, df + 2)
+    result = dc.HasNoX()(_add_n)(df, n=2)
+    tm.assert_frame_equal(result, df + 2)
+
+
 def test_has_no_nans():
     df = pd.DataFrame(np.random.randn(5, 3))
     result = ck.has_no_nans(df)
@@ -73,6 +84,39 @@ def test_has_no_nans():
     tm.assert_frame_equal(result, df + 2)
 
 
+def test_has_no_nones():
+    df = pd.DataFrame(np.random.randn(5, 3))
+    result = ck.has_no_nones(df)
+    tm.assert_frame_equal(df, result)
+
+    result = dc.HasNoNones()(_add_n)(df, 2)
+    tm.assert_frame_equal(result, df + 2)
+    result = dc.HasNoNones()(_add_n)(df, n=2)
+    tm.assert_frame_equal(result, df + 2)
+
+
+def test_has_no_infs():
+    df = pd.DataFrame(np.random.randn(5, 3))
+    result = ck.has_no_infs(df)
+    tm.assert_frame_equal(df, result)
+
+    result = dc.HasNoInfs()(_add_n)(df, 2)
+    tm.assert_frame_equal(result, df + 2)
+    result = dc.HasNoInfs()(_add_n)(df, n=2)
+    tm.assert_frame_equal(result, df + 2)
+
+
+def test_has_no_neg_infs():
+    df = pd.DataFrame(np.random.randn(5, 3))
+    result = ck.has_no_neg_infs(df)
+    tm.assert_frame_equal(df, result)
+
+    result = dc.HasNoNegInfs()(_add_n)(df, 2)
+    tm.assert_frame_equal(result, df + 2)
+    result = dc.HasNoNegInfs()(_add_n)(df, n=2)
+    tm.assert_frame_equal(result, df + 2)
+
+
 def test_has_no_nans_raises():
     df = pd.DataFrame(np.random.randn(5, 3))
     df.iloc[0, 0] = np.nan
@@ -81,6 +125,24 @@ def test_has_no_nans_raises():
 
     with pytest.raises(AssertionError):
         dc.HasNoNans()(_add_n)(df, n=2)
+
+
+def test_has_set_within_vals():
+    df = pd.DataFrame({'A': [1, 2, 3], 'B': ['a', 'b', 'c']})
+
+    items = {'A': [1, 2, 3], 'B': ['a', 'b', 'c']}
+    tm.assert_frame_equal(df, ck.has_set_within_vals(df, items))
+    tm.assert_frame_equal(df, dc.HasSetWithinVals(items=items)(_noop)(df))
+
+    items = {'A': [1, 2], 'B': ['a', 'b']}
+    tm.assert_frame_equal(df, ck.has_set_within_vals(df, items))
+    tm.assert_frame_equal(df, dc.HasSetWithinVals(items=items)(_noop)(df))
+
+    items = {'A': [1, 2, 4], 'B': ['a', 'b', 'd']}
+    with pytest.raises(AssertionError):
+        ck.has_set_within_vals(df, items)
+    with pytest.raises(AssertionError):
+        dc.HasSetWithinVals(items=items)(_noop)(df)
 
 
 def test_unique():